feat: add automatic database schema migration feature
- Introduced AUTO_MIGRATE_DB_SCHEMA environment variable to enable automatic database schema migration.
- Implemented migrateDbSchema function to handle the migration process.
- Updated main.ts to call migrateDbSchema during application bootstrap if the environment variable is set.
- Refactored sync-db-schema script to utilize migrateDbSchema for improved functionality.
- Updated .env and .env.example files to include the new environment variable.
