+++ apps/api/.env.development
@@ -4,6 +4,9 @@ PORT=5800
 WS_PORT=5801
 ORIGIN=http://localhost:5700,http://localhost:5173
 
+# Whether automatic database schema migration is enabled
+AUTO_MIGRATE_DB_SCHEMA=1
+
 # Static Endpoint Configuration
 STATIC_PUBLIC_ENDPOINT=
 STATIC_PRIVATE_ENDPOINT=
