+++ apps/api/src/main.ts
@@ -16,6 +16,7 @@ import { setTraceID } from './utils/middleware/set-trace-id';
 import { GlobalExceptionFilter } from './utils/filters/global-exception.filter';
 import { CustomWsAdapter } from './utils/adapters/ws-adapter';
 import { setupStatsig } from '@refly/telemetry-node';
+import { migrateDbSchema } from './utils/prisma';
 
 Sentry.init({
   dsn: process.env.SENTRY_DSN,
@@ -26,6 +27,11 @@ Sentry.init({
 });
 
 async function bootstrap() {
+  // Auto migrate db schema if the environment variable is set
+  if (process.env.AUTO_MIGRATE_DB_SCHEMA) {
+    await migrateDbSchema();
+  }
+
   const app = await NestFactory.create<NestExpressApplication>(AppModule, {
     rawBody: true,
     bufferLogs: false,
