+++ apps/api/src/scripts/sync-db-schema.ts
@@ -1,29 +1,10 @@
-import { execSync } from 'node:child_process';
-import { join, resolve } from 'node:path';
-import { existsSync } from 'node:fs';
-import { findNodeModules } from '../utils/runtime';
-
-// Start looking for node_modules from the directory of this script
-const nodeModulesPath =
-  findNodeModules(__dirname) ||
-  findNodeModules(resolve(process.cwd())) ||
-  join(process.cwd(), 'node_modules');
+import { migrateDbSchema } from '../utils/prisma';
 
 if (require.main === module) {
-  // Check if the prisma binary exists
-  let prismaBin = join(nodeModulesPath, '.bin', 'prisma');
+  console.log('Migrating DB schema...');
 
-  // Fallback to using pnpm bin if the above approach fails
-  if (!existsSync(prismaBin)) {
-    console.warn(
-      'Could not find Prisma binary using directory traversal, falling back to pnpm bin',
-    );
-    const binPath = execSync('pnpm bin', { encoding: 'utf-8' }).trim();
-    prismaBin = join(binPath, 'prisma');
-  }
-
-  execSync(
-    `${prismaBin} migrate diff --from-url ${process.env.DATABASE_URL} --to-schema-datamodel prisma/schema.prisma --script | ${prismaBin} db execute --stdin`,
-    { stdio: 'inherit' },
-  );
+  migrateDbSchema().then(() => {
+    console.log('DB schema migrated!');
+    process.exit(0);
+  });
 }
