+++ deploy/docker/docker-compose.yml
@@ -16,8 +16,6 @@ services:
     working_dir: /app/apps/api/dist
     extra_hosts:
       - "host.docker.internal:host-gateway"
-    command: >
-      sh -c "node scripts/sync-db-schema.js && node main.js"
     ports:
       - 5800:5800
       - 5801:5801
